Create new class IntArrayStack and delegate methods in DefaultCountingOutRhymer
